Hash the lost password and email confirm code/token in the database
Hash the lost password code/token in the database. This code is a Password Equivalent, so it MUST be hashed in case an attacker got his hands on your database. When a lost password code is requested, send the plaintext code to the user's email address, then hash it, save the hash in your database -- and throw away the original.