changed default value for caption field in file table
In SQL strict mode it throws an error and the record for the uploaded
file is not added, in consequences the parser will not render the file
correctly, because it assumes that file:image.png was never uploaded.

Field 'caption' doesn't have a default value

The upload handler do not set the files caption, only the filemeta
handler provides this option.